Quick bug fix for gradient map processor and others
Quick bug fix in graident map processor. Forgot to reset the UseMask to false if mask was null. Also forgot to hook up the OnInputRemoved for the mask, so as to reprocess the image when it is removed.

Updated curvature smooth and curvature in example graphs folder. Added gradient axial reflected, gradient axial, and perlin noise to examples and shelf. Removed all halftoneproc example, and replace with new one that is updated for the new function graph execute pin flow.

Fixed bug with parameter view not clearing when changing between function graphs and main graph etc. Also, should now clear when closing the tab for the graph.