Big changes...
All Primary Core graph and node functions moved out into its own project. The GL commands in the new Core project uses the new Materia.GLInterfaces from the Interfaces project. The GLInterfaces simply abstracts out all the used commands for the node graph, so OpenTK is not required for the core part specifically. Instead, any opengl .net library should be usable with it. There is also a DummyGL wrapper class that does nothing. It is there if you just want to load the graph and not actually generate images. The DummyGL allows you to simply make changes to the underlying graph settings if needed and etc without having to have a GL context / GL window.

Further abstraction will be required to handle multiple backend renderers such as Vulkan, Metal, and etc. That is a major todo for later, but for now the process has been slightly started with the first abstraction of the GL renderer

Fixed a bug in EmbossProcessor.cs where it was not using the proper variable for defining active texture position for opengl.

Various UI updates for pre-existing components.: Sliders now have an editable number input as well. Number input does not affect the slider position currently. The promote button grouping has been updated with a proper icon. It is faded out if not using a function. Faded in when a function is assigned to the property. Click the faded in button to access the function graph as usual. Dropdowns have been extended across the entire area instead of being right aligned only. Same for the boolean toggle buttons. Number inputs are now left aligned (except in the slider). Multirange slider has had its black slide button contrast increased to a slightly lighter color for easier viewing. Node headers have been updated to a blue color and white text.

Math nodes now have icons (not final) and previews for nodes that have user defined data.  Shelf now supports custom icon loading. To add an icon for a shelf item: Add a .png file to the Icons > Shelf folder with the same name as the .mtg file in the shelf folder (minus the .mtg extension).

Math3D are the files directly from OpenTK 3.0 github branch. I added it as a separate project, so the math in the Core project, RSMI project, etc would not rely on OpenTK directly for math only. The only projects that directly rely on OpenTK are Materia project and MTGRenderer project.