Some quick bug fixes....
Fixed a crash when opening an open mtg that did not have a section field on a GraphParameterValue. Forgot to check if it was null and set it to "Default".

Fixed a bug where if there was a comment or pin node in a function graph the function graph would not load properly. Because it would throw an exception that was ignored when a cast failed.