More gradient graphs + fixes.
More gradient graphs for shelf and examples folder. The gradient gaussian only works with this latest code. As the pow2 node was not working as expected previously.

Silly me had the pow2 node variable positions flipped on both CPU and GPU code. Pow2 was suppose to take the value and raise it by 2. However, it was doing the opposite where it was raising 2 by the specified value. It is now working as expected. Also note to self: add a regular pow node in where you can specifiy the amount to raise the value by.

Fixed where the 3d preview render stack image would show up in the 2D preview pane on graph open, close, and new.