Bug fixes...
Fixed bug where a graph could start at 0 zoom level.

Fixed a very serious bug, where when moving a panel, such as the shelf, could cause the current graph view to lose connections. This was due to a threading issue where it was still trying to load the ui for the primary graph when restoring the graph stack and breadcrumbs.

Fixed a bug where the graph stack / breadcrumbs were not being restored properly.

Fixed a bug where the 2D preview pane was still showing the 3D view. It would happen if you were already previewing a node but then you closed the graph.