Various bug fixes and reduction of redundent code / ui
Various bug fixes and reduction of redundent code / ui.
UI improvements / compaction for listing of properties in graph settings. Less vertical space is now taken up, for easier navigation.

Can now double click anywhere on an empty area in graph to bring up the graph settings.

Added editable dropdown for the various Get Float etc. It now displays any set variables in the dropdown that are available, including node, promoted, and custom parameters if available. Can still write out the name in the field as well. Arguments are also listed in the dropdown. It is filtered by the type of var you are getting, so if you are getting a float2 only float2 variables will be listed in the dropdown.

Simplified how Matrices will be handled. The underlying matrix that iwll be used for everything ia 4x4, since it can handle all possible transforms: scale, rotate, translate etc. Any vector that is not a float4 is converted temporarily to a float4 when used in multiplication against the matrix.