Improved Function Graph Speed
Improved function graph speed by caching order of nodes, and only recalculating if function graph nodes have changed.

Fixed a bug in FX node where luminosity random parameter getter was referencing Luminosity instead.

Added a switch node to atomic.

Fixed an issue where the Equals.mti should have actually been referencing EqualNode and not EqualsNode. Also renamed it to Equal.mti to correspond to the change.

Added a simple splatter graph based on tile generator. Added a weave generator graph.

Fixed a possible crash in function graph when a node was trying to be created, but caused a null reference exception if the node did not create properly.

Fixed negate node not properly updating its output type.

Fixed an issue where, when importing a function graph into another the current function graph name was overwritten with the imported function graph name.

Known Bugs:
Sometimes node graph lines are not properly deleted. When deleting a node that the output is connected to the input.

For some weird reason, when pasting a node and when connecting the output of it to another node, it does not properly connect in the underly UI / graph.