Fixed sluggishness of viewport updates with mouse movements
Fixed sluggishness of viewport updates with mouse movements, while still preventing gpu thrashing.

Will see GPU spike up for the following on the 3D View: when left mouse is down to rotate and when middle mouse is down for panning. This is a normal spike, as it turns on a constant 16.667ms loop for render updates while one of these two buttons are down.

Will see a gpu spike for the 2d preview when middle mouse button is down. This is a normal spike, as it turns on a constant 16.667ms loop for render updates while the button is down.

By doing the above, it provides a constant frame update for mouse movements, and prevents thrashing of GPU by not relying on mouse movement updates to update viewport. For instance, if just using mouse movement updates to update viewport, it is technically possible to thrash gpu up 90% depending on your mouse movement resolution.