Fixed issues relating to Bug #8, as well as an index out of bounds
Bug #8 is now fixed. Had to take into account several things, that got looked over. AKA forgot to dispose the previous UI node and its connections on the Node.OnInputRemoved event. Also adjusted the delegate for the Node.OnInputAdded and Node.OnOutputAdded. So, that the graphical UI had a chance to update the connections properly, and do a replace rather then a remove then add.

Also fixed an index out of range error for NodeOutput.Insert method.