Reworked UI rotation / scale stuff + HDR fixes
Reworked the UI stuff, because I realized that the scale + rotation in how it was applied was not going to work properly. Also, some other calculations were off when it came to child world positions etc. UI now operates as expected, with WorldMatrix simply being the overall rotation chain. Scale is now calculated as part of the AnchoredPosition. WorldSize now includes scale calculated into it.

Added in visibility check as well for Picking UI elements.

Fixed an issue where the mesh created via Scene.CreateMesh was not properly added to the transform heirachy.

Added new mesh modes for MeshRenderer.

Added in ActiveSkybox for Scene and made it so ActiveMeshes does not include the skybox.

Temporarily commented out uvArea in UI2DPreview for the moment.

Started on implementation of UI3DPreview window.

Removed unnecessary code from UIWindow. Also made sure content area could be raycast to.

Fixed issue in HdrFile loading. Now it reads all bytes into memory for fast  access and parallel scanline processing. Prior it would take over 1 second to read and process the hdr file incorrectly. It now processes a 4K hdr file in 120ms with 24 cores available.

Added a InvertedView accessor to Camera class. The InvertedView makes the camera become a orbital camera instead of the default FPS camera view.

fixed an issue in irradiance.glsl where it was using a non existent variable name for Normal.

Added back in a raw.glsl vertex shader for the ImageProcessor etc.

Corrected issue in MeshRenderer where it was trying to use the wrong dataSize for the VBO in some cases when calling DrawBasic(). DrawBasic() now properly disables the vertex attrib it does not need.

The Hdr Convereter.cs still needs further testing. And it is technically not fully finished yet.

Started reimplementation of HdriManagaer

Reworked the render passes classes.