Bug Fixes + 2D Preview Finished + More Stuff
Fixed an issue where the default Button text was not centered properly.

Fixed an issue where the BottomLeft / BottomHorizFill AnchorPosition was incorrectly calculated.

Fixed an issue on UI Picking where InVisiableArea() was still used, even if the underlying UIObject had a image that was using clipping. Clipping is now taken into account properly. If the UIObject has a UIDrawable that is clipped it now defaults back to the standard Contains(),

Added in actual light sphere to show where the light is in the 3D scene.
Added in most menu item activations  for 3D Preview. Still have to implement the major settings one via the Property Window.

Can now load custom meshes again for 3D preview.

All meshes are now available in the 3D preview menu.

Adjusted the uv.glsl vertex shader to be used as a ImageProcessor Adjusted UVrenderer as well. Added in a UVProcessor image processor.

2D preview is now complete and on parity with the previous WPF UI.

Added UV rendering to SceneStackRenderer

Fixed issues where I forgot to assign a name to FileDialogFilter.

Updated Avalonia to 0.10.4 to ensure that Materia is using the version that is properly signed and does not throw windows defender issues.

Fixed an issue in uniformcolor glsl where brightness check was still using lengthSqr instead of just length().

Added in a pbrbasic.glsl  vertex shader to use for PBRLight and PBRDepth

Fixed an issue in MeshRendere where I forgot to take into account the VertexAttributes in DrawBasic for Advanced Render Type.

Fixed an issue in BlurProcessor / MotionBlur Processor where the temp texture was not clamped to edge properly. Thus the blur could wrap, when it shouldn't.

Fixed and issue where ImageProcessor was not disabling CullFace and reenabling it on Complete(). Also now clears Stencil Buffer in Prepare()

Fixed an issue in BloomPass where DepthTest was not disabled / reenabled properly.

Added in FindComponent<T>() to UIObject to be able to find components that are a sublcass of another. E.g. find the first UIDrawable type, rather than haivng to check specific with HasComponent<UIImage>() for the target subclass.