Fixed several Avalonia UI related issues with dialogs etc
Fixed an issue where the color dialog was not working as expected and you could not change the hue via the mouse click and move.

Added support for reverse property value updates to some of the custom Avalonia controls such as NumberInput etc. So, if you click the 3d view with the light settings open and change position it will also update in the parameters / properties window if it is assigned.

Properties / Parameter window is now back and working.

Fixed an issue in custom Avalonia Slider where it wasn't updating the fill view properly after being added to the Avalonia window.

Added some margin / padding on the bottom right buttons in the 2D Preview window.

Added the rest of the menu actions to the 3D preview window.  The 3D preview window is now complete and on parity with the original WPF version.

Todo: Make it so you can toggle Skybox rendering or not on 3D preview.

Corrected the issue with the TextInput event on the Avalonia custom controls.

The Avalonia MainWindow will become the loading screen in essence.

Added some helpers to MVector for Vector -> MVector.

Added some missing methods from the new OpenTK Mathematics for Quaternion to our Quaternion class and etc.

Transform LocalEulerAngles now directly interfaces with Quaternion for getting and setting rotation.