Bug Fixes
Forgot to call function.AssignUniforms() after function.PrepareUniforms() for PixelProcessor node. They are now working again as expected.

Added a new draw event, and made adjustments to UIObject.SendMessage to take advantage of the draw event. Now, we get proper sibling clipping if one sibling is clipped and the other is not.

UIObjects now keep track of the parent that is in clip mode.

Added a helper to UIDrawable to check iif is in clip bounds.

Added IsInClipBounds() checks to UIImage and UIText. They will no longer render if they are fully outside the parent that is clipping them. Thus saving draw calls if they are not even in a viewable area.

Fixed issue with UIGraph grid offset. Forgot to multiply the offset by invZoom. Now, the grid properly offsets as expected and stays aligned with the things snapped to it. Also applied a 1.0f modulo to it so it will stay in -1.0 to 1.0 range.

Fixed an issue in UINodePoint where I forgot to check the node itself if it connected to an output and remove it, before attaching the new one in Connect().

Also removed the Threading for now in Graph. The only issue right now, is the changes kill GraphInstances within GraphInstances if they only schedule a single node, instead of using TryAndProcess() for the graph to schedule all.